server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
cloudinary==1.39.0
aiosqlite==0.20.0
python-dotenv==1.0.1